requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
dnspython==2.4.0
email-validator==2.0.0.post2
fastapi==0.100.0
frozenlist==1.4.0
gunicorn==21.1.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
idna==3.4
multidict==6.0.4
openai==0.27.8
packaging==23.1
pydantic==2.0.3
pydantic_core==2.3.0
pymongo==4.4.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
